+++ test/helpers/models.rb
@@ -66,3 +66,15 @@ end
 
 class StiRecordTypeA < StiRecord
 end
+
+class CustomMasterRecord < ActiveRecord::Base
+  include IdentityCache
+  has_many :custom_child_record, foreign_key: :master_id
+  self.primary_key = 'master_primary_key'
+end
+
+class CustomChildRecord < ActiveRecord::Base
+  include IdentityCache
+  belongs_to :custom_master_record, foreign_key: :master_id
+  self.primary_key = 'child_primary_key'
+end